refactor: 유저 정보 불러오는 코드 커스텀 훅으로 분리
